backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6  # 文件上传支持

# 数据库
sqlalchemy==2.0.23

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP请求
requests==2.31.0
aiohttp==3.9.1  # 异步HTTP（可选）

# 爬虫与解析
beautifulsoup4==4.12.2
lxml==4.9.3
html2text==2020.1.16

# 视频字幕
yt-dlp==2023.11.16

# Markdown处理
markdown==3.5.1
pymdown-extensions==10.5

# 任务调度
apscheduler==3.10.4

# 工具库
python-dotenv==1.0.0  # 环境变量
pydantic==2.5.0  # 数据验证
pydantic-settings==2.1.0  # 配置管理

# 文件处理（可选，Windows需额外安装）
# python-magic==0.4.27  # 文件类型检测

# 日志（使用标准logging，不需要loguru）
# loguru==0.7.2
